Refactor: re-organize shared components
