+++ app/src/main/java/com/github/meeplemeet/ui/components/BusinessComponents.kt
@@ -10,11 +10,8 @@ import android.content.Context
 import android.text.format.DateFormat
 import android.widget.Toast
 import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -30,8 +27,6 @@ import androidx.compose.material.icons.filled.Email
 import androidx.compose.material.icons.filled.Language
 import androidx.compose.material.icons.filled.Phone
 import androidx.compose.material.icons.filled.Place
-import androidx.compose.material.icons.filled.Remove
-import androidx.compose.material.icons.filled.VideogameAsset
 import androidx.compose.material.icons.outlined.Edit
 import androidx.compose.material3.*
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -55,12 +50,7 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import com.github.meeplemeet.model.account.Account
-import com.github.meeplemeet.model.shared.GameUIState
-import com.github.meeplemeet.model.shared.game.Game
 import com.github.meeplemeet.model.shops.OpeningHours
-import com.github.meeplemeet.model.shops.Shop
-import com.github.meeplemeet.model.shops.ShopSearchViewModel
 import com.github.meeplemeet.model.shops.TimeSlot
 import com.github.meeplemeet.ui.FocusableInputField
 import com.github.meeplemeet.ui.space_renter.SpaceRenterTestTags
@@ -971,380 +961,6 @@ fun ActionBar(
                     Icon(Icons.Filled.Check, contentDescription = null)
                     Spacer(Modifier.width(Dimensions.Spacing.medium))
                     Text(primaryButtonText)
-                  }
-            }
-      }
-}
-
-/* =============================================================================
- * Game search + quantity
- * ============================================================================= */
-
-/**
- * A composable function that displays a quantity input with +/- buttons and a label.
- *
- * @param value The current quantity value.
- * @param onValueChange A callback function that is invoked when the quantity value changes.
- * @param range The range of valid quantity values.
- * @param modifier The modifier to be applied to the quantity input.
- */
-@Composable
-fun GameAddUI(value: Int, onValueChange: (Int) -> Unit, modifier: Modifier = Modifier) {
-  Column(modifier.testTag(ShopComponentsTestTags.QTY_CONTAINER)) {
-    Text(
-        ShopUiDefaults.StringsMagicNumbers.QUANTITY,
-        style = MaterialTheme.typography.headlineSmall,
-        modifier = Modifier.testTag(ShopComponentsTestTags.QTY_LABEL))
-
-    Spacer(Modifier.height(Dimensions.Spacing.large))
-
-    Row(
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(Dimensions.Spacing.medium),
-        modifier = Modifier.fillMaxWidth()) {
-          IconButton(
-              onClick = {
-                val newValue = (value - 1)
-                onValueChange(newValue)
-              },
-              enabled = value > 0,
-              modifier = Modifier.testTag(ShopComponentsTestTags.QTY_MINUS_BUTTON)) {
-                Icon(Icons.Filled.Remove, contentDescription = "Decrease quantity")
-              }
-
-          FocusableInputField(
-              value = value.toString(),
-              onValueChange = { newText ->
-                val newValue = newText.toIntOrNull() ?: 0
-                if (newValue > 0) {
-                  onValueChange(newValue)
-                }
-              },
-              keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
-              singleLine = true,
-              textStyle = MaterialTheme.typography.bodyLarge.copy(textAlign = TextAlign.Center),
-              modifier = Modifier.weight(1f).testTag(ShopComponentsTestTags.QTY_INPUT_FIELD))
-
-          IconButton(
-              onClick = {
-                val newValue = (value + 1)
-                onValueChange(newValue)
-              },
-              enabled = true,
-              modifier = Modifier.testTag(ShopComponentsTestTags.QTY_PLUS_BUTTON)) {
-                Icon(Icons.Filled.Add, contentDescription = "Increase quantity")
-              }
-        }
-  }
-}
-
-/* =============================================================================
- * Game stock dialog
- * ============================================================================= */
-
-/**
- * A composable function that displays a dialog for adding a game to stock with search and quantity
- * selection.
- *
- * @param onQueryChange A callback function that is invoked when the search query changes.
- * @param quantity The current quantity value.
- * @param onQuantityChange A callback function that is invoked when the quantity value changes.
- * @param existingIds A set of existing game IDs to prevent duplicates.
- * @param onDismiss A callback function that is invoked when the dialog is dismissed.
- * @param onSave A callback function that is invoked when the save button is clicked.
- */
-@Composable
-fun GameStockDialog(
-    owner: Account,
-    shop: Shop?,
-    viewModel: ShopSearchViewModel,
-    gameUIState: GameUIState,
-    onQueryChange: (String) -> Unit,
-    quantity: Int,
-    onQuantityChange: (Int) -> Unit,
-    existingIds: Set<String> = emptySet(),
-    onDismiss: () -> Unit,
-    onSave: () -> Unit,
-) {
-  val selectedGame = gameUIState.fetchedGame
-  val isDuplicate = selectedGame?.uid?.let { it in existingIds } ?: false
-
-  AlertDialog(
-      onDismissRequest = onDismiss,
-      shape = MaterialTheme.shapes.extraLarge,
-      containerColor = AppColors.primary,
-      title = {
-        Box(
-            Modifier.fillMaxWidth()
-                .testTag(ShopComponentsTestTags.GAME_DIALOG_TITLE)
-                .background(AppColors.primary),
-            contentAlignment = Alignment.Center) {
-              Text(
-                  ShopUiDefaults.StringsMagicNumbers.GAME_DIALOG_TITLE,
-                  style = MaterialTheme.typography.headlineSmall)
-            }
-      },
-      text = {
-        Column(Modifier.fillMaxWidth().testTag(ShopComponentsTestTags.GAME_DIALOG_BODY)) {
-          ShopGameSearchBar(
-              owner,
-              shop,
-              viewModel,
-              gameUIState.fetchedGame,
-              existingIds,
-              inputFieldTestTag = ShopComponentsTestTags.GAME_SEARCH_FIELD,
-              dropdownItemTestTag = ShopComponentsTestTags.GAME_SEARCH_ITEM)
-
-          if (isDuplicate) {
-            Spacer(Modifier.height(Dimensions.Padding.mediumSmall))
-            Text(
-                ShopUiDefaults.StringsMagicNumbers.DUPLICATE_GAME,
-                color = MaterialTheme.colorScheme.error,
-                style = MaterialTheme.typography.bodySmall,
-                modifier = Modifier.testTag(ShopComponentsTestTags.GAME_DIALOG_HELPER))
-          }
-
-          Spacer(Modifier.height(Dimensions.Spacing.extraLarge))
-
-          GameAddUI(
-              value = quantity,
-              onValueChange = onQuantityChange,
-              modifier =
-                  Modifier.testTag(ShopComponentsTestTags.GAME_DIALOG_SLIDER)
-                      .background(AppColors.primary))
-        }
-      },
-      dismissButton = {
-        TextButton(
-            onClick = onDismiss,
-            modifier = Modifier.testTag(ShopComponentsTestTags.GAME_DIALOG_CANCEL)) {
-              Text(ShopUiDefaults.StringsMagicNumbers.BTN_CANCEL)
-            }
-      },
-      confirmButton = {
-        TextButton(
-            onClick = onSave,
-            enabled = gameUIState.fetchedGame != null && !isDuplicate && quantity > 0,
-            modifier = Modifier.testTag(ShopComponentsTestTags.GAME_DIALOG_SAVE)) {
-              Text(ShopUiDefaults.StringsMagicNumbers.BTN_SAVE)
-            }
-      })
-}
-
-/* =============================================================================
- * Games: grid + item (with optional delete)
- * ============================================================================= */
-
-/**
- * A composable function that displays a section of games in a grid or list format, with optional
- * title and delete buttons.
- *
- * @param games A list of pairs containing [Game] objects and their corresponding quantities.
- * @param modifier The modifier to be applied to the game list section.
- * @param clickableGames A boolean indicating whether the game items are clickable.
- * @param title An optional title for the game list section.
- * @param hasDeleteButton A boolean indicating whether the game items have delete buttons.
- * @param onClick A callback function that is invoked when a game item is clicked.
- * @param onDelete A callback function that is invoked when a game item is deleted.
- */
-@Composable
-fun GameListSection(
-    games: List<Pair<Game, Int>>,
-    modifier: Modifier = Modifier,
-    clickableGames: Boolean = false,
-    title: String? = null,
-    hasDeleteButton: Boolean = false,
-    onClick: (Game) -> Unit = {},
-    onDelete: (Game) -> Unit = {},
-) {
-  Column(
-      verticalArrangement = Arrangement.spacedBy(Dimensions.Spacing.medium),
-      modifier = modifier.fillMaxWidth()) {
-        if (title != null) {
-          Text(
-              title,
-              style = MaterialTheme.typography.titleLarge,
-              fontWeight = FontWeight.SemiBold,
-          )
-        }
-
-        LazyColumn(
-            verticalArrangement = Arrangement.spacedBy(Dimensions.Spacing.medium),
-            contentPadding = PaddingValues(bottom = Dimensions.Spacing.extraLarge),
-            modifier = Modifier.heightIn(max = Dimensions.ContainerSize.maxListHeight)) {
-              items(items = games, key = { it.first.uid }) { (game, count) ->
-                GameItem(
-                    game = game,
-                    count = count,
-                    clickable = clickableGames,
-                    onClick = onClick,
-                    hasDeleteButton = hasDeleteButton,
-                    onDelete = onDelete)
-              }
-            }
-      }
-}
-
-/**
- * A composable function that displays a game item with its name, icon, quantity badge, and optional
- * delete button.
- *
- * @param game The [Game] object to be displayed.
- * @param count The quantity of the game.
- * @param modifier The modifier to be applied to the game item.
- * @param clickable A boolean indicating whether the game item is clickable.
- * @param onClick A callback function that is invoked when the game item is clicked.
- * @param hasDeleteButton A boolean indicating whether the game item has a delete button.
- * @param onDelete A callback function that is invoked when the delete button is clicked.
- */
-@Composable
-fun GameItem(
-    game: Game,
-    count: Int,
-    modifier: Modifier = Modifier,
-    clickable: Boolean = false,
-    onClick: (Game) -> Unit = {},
-    hasDeleteButton: Boolean = false,
-    onDelete: (Game) -> Unit = {},
-) {
-  Card(
-      modifier =
-          modifier
-              .fillMaxWidth()
-              .testTag("${ShopComponentsTestTags.SHOP_GAME_PREFIX}${game.uid}")
-              .clickable(enabled = clickable, onClick = { onClick(game) }),
-      colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surface)) {
-        Row(
-            modifier = Modifier.padding(Dimensions.Padding.medium),
-            verticalAlignment = Alignment.CenterVertically) {
-
-              // Icon + badge
-              Box(
-                  modifier = Modifier.size(Dimensions.IconSize.huge),
-                  contentAlignment = Alignment.Center) {
-                    BadgedBox(
-                        badge = {
-                          // Only show badge if count > 0
-                          if (count > 0) {
-                            val max = ShopUiDefaults.RangesMagicNumbers.qtyGameDialog.last
-                            val label = if (count > max) "$max+" else count.toString()
-                            Badge(
-                                modifier =
-                                    Modifier.offset(
-                                            x = Dimensions.BadgeSize.offsetX,
-                                            y = Dimensions.BadgeSize.offsetY)
-                                        .defaultMinSize(
-                                            minWidth = Dimensions.BadgeSize.minSize,
-                                            minHeight = Dimensions.BadgeSize.minSize),
-                                containerColor = MaterialTheme.colorScheme.inversePrimary) {
-                                  Text(
-                                      label,
-                                      style = MaterialTheme.typography.labelSmall,
-                                      maxLines = Dimensions.Numbers.singleLine,
-                                      softWrap = false,
-                                      modifier =
-                                          Modifier.padding(horizontal = Dimensions.Spacing.small))
-                                }
-                          }
-                        }) {
-                          Icon(Icons.Filled.VideogameAsset, contentDescription = null)
-                        }
-                  }
-
-              Spacer(Modifier.width(Dimensions.Spacing.medium))
-
-              // Name centered
-              Column(
-                  modifier = Modifier.weight(1f),
-                  horizontalAlignment = Alignment.CenterHorizontally,
-                  verticalArrangement = Arrangement.Center) {
-                    Text(
-                        game.name,
-                        fontWeight = FontWeight.SemiBold,
-                        modifier = Modifier.fillMaxWidth(),
-                        textAlign = TextAlign.Center)
-                  }
-
-              // Optional delete
-              if (hasDeleteButton) {
-                IconButton(
-                    onClick = { onDelete(game) },
-                    modifier =
-                        Modifier.testTag("${ShopComponentsTestTags.SHOP_GAME_DELETE}:${game.uid}"),
-                    colors =
-                        IconButtonDefaults.iconButtonColors(
-                            contentColor = MaterialTheme.colorScheme.error)) {
-                      Icon(
-                          Icons.Filled.Delete, contentDescription = "Remove ${game.name} from list")
-                    }
-              }
-            }
-      }
-}
-
-/* =============================================================================
- * Games: editable item helper
- * ============================================================================= */
-
-/** Editable game row used in edit screen (inline quantity +/- and delete). */
-@Composable
-fun EditableGameItem(
-    game: Game,
-    count: Int,
-    onQuantityChange: (Game, Int) -> Unit,
-    onDelete: (Game) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-  Card(
-      modifier =
-          modifier.fillMaxWidth().testTag("${ShopComponentsTestTags.SHOP_GAME_PREFIX}${game.uid}"),
-      colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.surface)) {
-        Row(
-            Modifier.padding(Dimensions.Padding.medium),
-            verticalAlignment = Alignment.CenterVertically) {
-              Text(
-                  game.name,
-                  style = MaterialTheme.typography.bodyLarge,
-                  modifier = Modifier.weight(1f))
-              Row(
-                  verticalAlignment = Alignment.CenterVertically,
-                  horizontalArrangement = Arrangement.spacedBy(Dimensions.Spacing.small)) {
-                    IconButton(
-                        onClick = { onQuantityChange(game, (count - 1).coerceAtLeast(0)) },
-                        enabled = count > 0,
-                        modifier =
-                            Modifier.testTag(ShopComponentsTestTags.SHOP_GAME_MINUS_BUTTON)) {
-                          Icon(Icons.Filled.Remove, contentDescription = "Decrease quantity")
-                        }
-                    FocusableInputField(
-                        value = count.toString(),
-                        onValueChange = { newText ->
-                          val newValue = newText.toIntOrNull() ?: 0
-                          if (newValue >= 0) {
-                            onQuantityChange(game, newValue)
-                          }
-                        },
-                        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
-                        singleLine = true,
-                        textStyle =
-                            MaterialTheme.typography.bodyLarge.copy(textAlign = TextAlign.Center),
-                        modifier =
-                            Modifier.width(Dimensions.ComponentWidth.inputFieldWidth)
-                                .testTag(ShopComponentsTestTags.SHOP_GAME_QTY_INPUT))
-                    IconButton(
-                        onClick = { onQuantityChange(game, count + 1) },
-                        modifier = Modifier.testTag(ShopComponentsTestTags.SHOP_GAME_PLUS_BUTTON)) {
-                          Icon(Icons.Filled.Add, contentDescription = "Increase quantity")
-                        }
-                  }
-              IconButton(
-                  onClick = { onDelete(game) },
-                  colors =
-                      IconButtonDefaults.iconButtonColors(
-                          contentColor = MaterialTheme.colorScheme.error),
-                  modifier =
-                      Modifier.testTag("${ShopComponentsTestTags.SHOP_GAME_DELETE}:${game.uid}")) {
-                    Icon(Icons.Filled.Delete, contentDescription = "Delete game")
                   }
             }
       }
